Multi-machine refactor phase 6 (part 6): split public machine_profile.h + layering check
Realize the proposal §4.3 layering split:
- New src/core/machine_profile.h holds the PUBLIC descriptor + capability
  types (floppy/scsi slots, floppy_kind_t, mmu_kind_t, hw_profile_t) and the
  registry API (machine_find / machine_list / *_to_string). This is the one
  machine-subsystem header the platform-agnostic core may include.
- src/machines/machine.h is now a thin umbrella: it includes machine_profile.h
  and adds only the implementation-side surface (the extern profile objects).
- The two core includers (system.c, system_config.h) include machine_profile.h
  instead of machine.h — core no longer reaches into src/machines/ for types.
- New tests/integration/core-layering enforces the rule in CI: no file under
  src/core/ may #include any header that lives under src/machines/ (it builds
  the machine-header set dynamically, so it stays correct as files move).

Verified: headless + WASM build clean; core-layering check passes.
