+++ tests/integration/core-layering/test.script
@@ -1,0 +1,3 @@
+# Real logic is in run.sh (a static grep over src/core).  The harness requires
+# a test.script to exist; this stub is unused by the TEST_RUNNER path.
+quit
